requirements.txt
fastapi
uvicorn
sqlalchemy
psycopg2-binary
python-jose[cryptography]
passlib[bcrypt]
pydantic
python-multipart
langchain
faiss-cpu
openai
PyPDF2
streamlit
requests
python-dotenv
